refactor: extract location note helpers to utils file
Move safeDisplay, formatLocationLink, buildAssetListMarkup,
and buildKitListMarkup from service.server.ts to a dedicated
utils.ts file. Service files should only contain functions
that interact with the DB model.
